move DatasetsController code into modules
