+++ app/controllers/concerns/deepblue/visibility_change_controller_behavior.rb
@@ -1,0 +1,64 @@
+# frozen_string_literal: true
+
+module Deepblue
+  module VisibilityChangeControllerBehavior
+    extend ActiveSupport::Concern
+
+    included do
+      before_action :visiblity_changed,            only: [:update]
+      after_action :visibility_changed_update,     only: [:update]
+    end
+
+    def visiblity_changed
+      # ::Deepblue::LoggingHelper.bold_debug [ Deepblue::LoggingHelper.here,
+      #                                        Deepblue::LoggingHelper.called_from,
+      #                                        Deepblue::LoggingHelper.obj_class( 'class', self ),
+      #                                        "" ]
+      if visibility_to_private?
+        mark_as_set_to_private
+      elsif visibility_to_public?
+        mark_as_set_to_public
+      end
+    end
+
+    def visibility_changed_update
+      # ::Deepblue::LoggingHelper.bold_debug [ Deepblue::LoggingHelper.here,
+      #                                        Deepblue::LoggingHelper.called_from,
+      #                                        Deepblue::LoggingHelper.obj_class( 'class', self ),
+      #                                        "" ]
+      if curation_concern.private? && @visibility_changed_to_private
+       workflow_unpublish
+      elsif curation_concern.public? && @visibility_changed_to_public
+        workflow_publish
+      end
+    end
+
+    def visibility_to_private?
+      # ::Deepblue::LoggingHelper.bold_debug [ Deepblue::LoggingHelper.here,
+      #                                        Deepblue::LoggingHelper.called_from,
+      #                                        Deepblue::LoggingHelper.obj_class( 'class', self ),
+      #                                        "" ]
+      return false if curation_concern.private?
+      params[params_key]['visibility'] == Hydra::AccessControls::AccessRight::VISIBILITY_TEXT_VALUE_PRIVATE
+    end
+
+    def visibility_to_public?
+      # ::Deepblue::LoggingHelper.bold_debug [ Deepblue::LoggingHelper.here,
+      #                                        Deepblue::LoggingHelper.called_from,
+      #                                        Deepblue::LoggingHelper.obj_class( 'class', self ),
+      #                                        "" ]
+      return false if curation_concern.public?
+      params[PARAMS_KEY]['visibility'] == Hydra::AccessControls::AccessRight::VISIBILITY_TEXT_VALUE_PUBLIC
+    end
+
+    def mark_as_set_to_private
+      @visibility_changed_to_public = false
+      @visibility_changed_to_private = true
+    end
+
+    def mark_as_set_to_public
+      @visibility_changed_to_public = true
+      @visibility_changed_to_private = false
+    end
+  end
+end
